+++ persistence/unit_test.go
@@ -7,13 +7,13 @@ import (
 
 type testProductInterface interface {
 	data() int
-	ID() uint
+	ID() uint64
 }
 
 type emptyProductInterface struct {
 }
 
-func (e emptyProductInterface) ID() uint {
+func (e emptyProductInterface) ID() uint64 {
 	return 0
 }
 
@@ -21,7 +21,7 @@ type testIntType struct {
 	u int
 }
 
-func (t testIntType) ID() uint { return uint(t.u) }
+func (t testIntType) ID() uint64 { return uint64(t.u) }
 
 func (t testIntType) data() int { return t.u }
 
@@ -29,7 +29,7 @@ type testByteType struct {
 	u byte
 }
 
-func (t testByteType) ID() uint { return uint(t.u) }
+func (t testByteType) ID() uint64 { return uint64(t.u) }
 
 func (t testByteType) data() int { return int(t.u) }
 
@@ -37,7 +37,7 @@ type testRuneType struct {
 	u rune
 }
 
-func (t testRuneType) ID() uint { return uint(t.u) }
+func (t testRuneType) ID() uint64 { return uint64(t.u) }
 
 func (t testRuneType) data() int { return int(t.u) }
