+++ protocol/websocket/message/initialize.go
@@ -3,9 +3,9 @@ package message
 import "github.com/sunist-c/genius-invokation-simulator-backend/enum"
 
 type MatchingMessage struct {
-	UID        uint
-	Characters []uint
-	CardDeck   []uint
+	UID        uint64
+	Characters []uint64
+	CardDeck   []uint64
 }
 
 type GameOptions struct {
